requirements.txt
# ── Core Framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Database ────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
asyncpg==0.29.0

# ── Redis ───────────────────────────────────────────────────────
redis==5.0.4

# ── Kafka ───────────────────────────────────────────────────────
confluent-kafka==2.4.0

# ── ML — Core ───────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.0

# ── ML — Models ─────────────────────────────────────────────────
lightgbm==4.3.0
torch==2.3.0

# ── ML — Imbalance ──────────────────────────────────────────────
imbalanced-learn==0.12.2

# ── ML — Explainability ─────────────────────────────────────────
shap==0.45.1

# ── ML — Experiment Tracking ────────────────────────────────────
mlflow==2.12.1

# ── ML — Drift Monitoring ───────────────────────────────────────
evidently==0.4.30

# ── Task Queue ──────────────────────────────────────────────────
celery==5.3.6

# ── Template Engine ─────────────────────────────────────────────
Jinja2==3.1.4

# ── Data Generation ─────────────────────────────────────────────
faker==24.9.0

# ── API & HTTP ──────────────────────────────────────────────────
httpx==0.27.0
requests==2.32.3

# ── Serialization ───────────────────────────────────────────────
fastavro==1.9.4

# ── Logging ─────────────────────────────────────────────────────
structlog==24.1.0

# ── Metrics ─────────────────────────────────────────────────────
prometheus-client==0.20.0

# ── Utilities ───────────────────────────────────────────────────
joblib==1.4.0
tqdm==4.66.2
python-dateutil==2.9.0
pytz==2024.1